backend/requirements.txt
fastapi>=0.110.0
uvicorn>=0.27.0
python-pptx>=1.0.0
pydantic>=2.6.0
matplotlib>=3.8.0
